+++ src/frontend/apps/impress/src/features/docs/doc-management/api/useRemoveDoc.tsx
@@ -33,19 +33,19 @@ export const useRemoveDoc = ({
   return useMutation<void, APIError, RemoveDocProps>({
     mutationFn: removeDoc,
     ...options,
-    onSuccess: (data, variables, context) => {
+    onSuccess: (data, variables, onMutateResult, context) => {
       listInvalidQueries?.forEach((queryKey) => {
         void queryClient.invalidateQueries({
           queryKey: [queryKey],
         });
       });
       if (options?.onSuccess) {
-        void options.onSuccess(data, variables, context);
+        void options.onSuccess(data, variables, onMutateResult, context);
       }
     },
-    onError: (error, variables, context) => {
+    onError: (error, variables, onMutateResult, context) => {
       if (options?.onError) {
-        void options.onError(error, variables, context);
+        void options.onError(error, variables, onMutateResult, context);
       }
     },
   });